+++ packages/features/bookings/lib/handleNewBooking/getOriginalRescheduledBooking.ts
@@ -1,59 +1,14 @@
 import type { Prisma } from "@prisma/client";
 
-import prisma from "@calcom/prisma";
-import { BookingStatus } from "@calcom/prisma/enums";
+import { BookingRepository } from "@calcom/lib/server/repository/booking";
+
+import { validateOriginalRescheduledBooking } from "./validateOriginalRescheduledBooking";
 
 export async function getOriginalRescheduledBooking(uid: string, seatsEventType?: boolean) {
-  return prisma.booking.findFirst({
-    where: {
-      uid: uid,
-      status: {
-        in: [BookingStatus.ACCEPTED, BookingStatus.CANCELLED, BookingStatus.PENDING],
-      },
-    },
-    include: {
-      attendees: {
-        select: {
-          name: true,
-          email: true,
-          locale: true,
-          timeZone: true,
-          phoneNumber: true,
-          ...(seatsEventType && { bookingSeat: true, id: true }),
-        },
-      },
-      user: {
-        select: {
-          id: true,
-          name: true,
-          email: true,
-          locale: true,
-          timeZone: true,
-          destinationCalendar: true,
-          credentials: {
-            select: {
-              id: true,
-              userId: true,
-              key: true,
-              type: true,
-              teamId: true,
-              appId: true,
-              invalid: true,
-              user: {
-                select: {
-                  email: true,
-                },
-              },
-            },
-          },
-        },
-      },
-      destinationCalendar: true,
-      payment: true,
-      references: true,
-      workflowReminders: true,
-    },
-  });
+  const originalBooking = await BookingRepository.findOriginalRescheduledBooking(uid, seatsEventType);
+  validateOriginalRescheduledBooking(originalBooking);
+
+  return originalBooking;
 }
 
 export type BookingType = Prisma.PromiseReturnType<typeof getOriginalRescheduledBooking>;
