+++ packages/lib/server/repository/booking.ts
@@ -336,4 +336,57 @@ export class BookingRepository {
 
     return [...collectiveRoundRobinBookingsOwner, ...collectiveRoundRobinBookingsAttendee];
   }
+
+  static async findOriginalRescheduledBooking(uid: string, seatsEventType?: boolean) {
+    return await prisma.booking.findFirst({
+      where: {
+        uid: uid,
+        status: {
+          in: [BookingStatus.ACCEPTED, BookingStatus.CANCELLED, BookingStatus.PENDING],
+        },
+      },
+      include: {
+        attendees: {
+          select: {
+            name: true,
+            email: true,
+            locale: true,
+            timeZone: true,
+            phoneNumber: true,
+            ...(seatsEventType && { bookingSeat: true, id: true }),
+          },
+        },
+        user: {
+          select: {
+            id: true,
+            name: true,
+            email: true,
+            locale: true,
+            timeZone: true,
+            destinationCalendar: true,
+            credentials: {
+              select: {
+                id: true,
+                userId: true,
+                key: true,
+                type: true,
+                teamId: true,
+                appId: true,
+                invalid: true,
+                user: {
+                  select: {
+                    email: true,
+                  },
+                },
+              },
+            },
+          },
+        },
+        destinationCalendar: true,
+        payment: true,
+        references: true,
+        workflowReminders: true,
+      },
+    });
+  }
 }
